refactor create and join conference
